+++ src/components/LoadingMask.tsx
@@ -2,24 +2,27 @@
 import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import { LoadingOutlined } from '@ant-design/icons';
-import { Spin } from 'antd';
+import { Spin, Typography } from 'antd';
 
 interface Props {
 	className?: string;
 	visible: boolean;
 	delay?: number;
 	fullscreen?: boolean;
+	opacity?: number;
+	message?: string;
 }
 
 interface ContainerProps {
 	fullscreen?: boolean;
+	opacity?: number;
 }
 
-const StyledSpin = styled(Spin)`
+const StyledSpinContainer = styled.div`
 	position: absolute;
 	left: 50%;
 	top: 50%;
-	transform: translate(0, -50%);
+	transform: translate(-50%, -50%);
 `;
 
 const Container = styled.div<ContainerProps>`
@@ -28,9 +31,16 @@ const Container = styled.div<ContainerProps>`
 	left: 0;
 	right: 0;
 	bottom: 0;
-	z-index: 1000;
+	z-index: 1001;
 	height: 100%;
-	background-color: rgba(0, 0, 0, 0.7);
+	background-color: ${(props): string => (props.opacity === undefined ? 'rgba(0, 0, 0, 0.7)' : `rgba(0, 0, 0, ${props.opacity})`)};
+`;
+
+const StyledText = styled(Typography.Text)`
+	position: absolute;
+	left: 50%;
+	top: calc(50% + 64px);
+	transform: translate(-50%, -50%);
 `;
 
 const LoadingMask: React.FunctionComponent<Props> = (props: Props) => {
@@ -54,8 +64,11 @@ const LoadingMask: React.FunctionComponent<Props> = (props: Props) => {
 
 	const renderMask = (): JSX.Element => {
 		return (
-			<Container className={props.className} fullscreen={props.fullscreen}>
-				<StyledSpin indicator={<LoadingOutlined style={{ fontSize: '64px' }} />} />
+			<Container className={props.className} fullscreen={props.fullscreen} opacity={props.opacity}>
+				<StyledSpinContainer>
+					<Spin indicator={<LoadingOutlined style={{ fontSize: '64px' }} />} />
+				</StyledSpinContainer>
+				{props.message && <StyledText>{props.message}</StyledText>}
 			</Container>
 		);
 	};