+++ test/helpers/window.mock.ts
@@ -6,3 +6,9 @@ export const ResizeObserver = class {
 	observe = observe;
 	disconnect = disconnect;
 };
+export const log = {
+	info: jest.fn(),
+	debug: jest.fn(),
+	error: jest.fn(),
+	warn: jest.fn(),
+};