+++ test/store/settings.test.ts
@@ -4,8 +4,17 @@ import reducer, { actions, initialState } from '../../src/store/settings';
 import { thunks } from '../../src/store/';
 import { Settings } from '../../src/store/types';
 import { createAction, mSettings } from '../helpers/test.helper';
+jest.mock('../../src/types/components', () => {
+	return {
+		openNotificationWithIcon: jest.fn(),
+	};
+});
+import { openNotificationWithIcon } from '../../src/types/components';
 
 describe('store/settings', () => {
+	beforeEach(() => {
+		jest.clearAllMocks();
+	});
 	it('Sets API key', () => {
 		// given
 		const key = 'api_key';
@@ -128,5 +137,94 @@ describe('store/settings', () => {
 
 		// then
 		expect(result).toStrictEqual({ ...settings, dashboard: initialState.dashboard });
+	});
+	it('Updates whole state when loadSettings is fulfilled and settings exist', () => {
+		// given
+		const settings: Partial<Settings> = {
+			...mSettings({
+				apiKey: 'asdasfasf',
+				gelbooruUsername: 'asdasd',
+				imagesFolderPath: 'aasdasdasdff',
+				theme: 'light',
+			}),
+			dashboard: undefined,
+		};
+		const action = createAction(thunks.settings.loadSettings.fulfilled.type, settings);
+
+		// when
+		const result = reducer(undefined, action);
+
+		// then
+		expect(result).toStrictEqual({ ...settings, dashboard: initialState.dashboard });
+	});
+	it('Opens notification when importDatabase is fulfilled', () => {
+		// given
+		const action = createAction(thunks.settings.importDatabase.fulfilled.type, true);
+
+		// when
+		reducer(undefined, action);
+
+		// then
+		expect(openNotificationWithIcon).toBeCalledWith('success', 'Import finished', 'Database was succesfully restored!');
+	});
+	it('Does not open notification when importDatabase dialog is closed without selecing a file', () => {
+		// given
+		const action = createAction(thunks.settings.importDatabase.fulfilled.type, false);
+
+		// when
+		reducer(undefined, action);
+
+		// then
+		expect(openNotificationWithIcon).toBeCalledTimes(0);
+	});
+	it('Opens notification when importDatabase is rejected', () => {
+		// given
+		const action = createAction(thunks.settings.importDatabase.rejected.type, true);
+
+		// when
+		reducer(undefined, action);
+
+		// then
+		expect(openNotificationWithIcon).toBeCalledWith(
+			'error',
+			'Could not import database',
+			'Error occured while trying to restore backup.',
+			5
+		);
+	});
+	it('Opens notification when exportDatabase is fulfilled', () => {
+		// given
+		const action = createAction(thunks.settings.exportDatabase.fulfilled.type, true);
+
+		// when
+		reducer(undefined, action);
+
+		// then
+		expect(openNotificationWithIcon).toBeCalledWith('success', 'Export finished', 'Database was succesfully exported!');
+	});
+	it('Does not open notification when exportDatabase dialog is closed without selecing a file', () => {
+		// given
+		const action = createAction(thunks.settings.exportDatabase.fulfilled.type, false);
+
+		// when
+		reducer(undefined, action);
+
+		// then
+		expect(openNotificationWithIcon).toBeCalledTimes(0);
+	});
+	it('Opens notification when exportDatabase is rejected', () => {
+		// given
+		const action = createAction(thunks.settings.exportDatabase.rejected.type, true);
+
+		// when
+		reducer(undefined, action);
+
+		// then
+		expect(openNotificationWithIcon).toBeCalledWith(
+			'error',
+			'Could not export database',
+			'Error occured while trying to create a database backup.',
+			5
+		);
 	});
 });